# apps/server/internal/docs/swagger.yaml
basePath: /
definitions:
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: Internal admin platform API documentation.
  title: Admin Service API
  version: 0.1.0
paths:
...
  /v1/monitor/logs/login:
    get:
      description: 按用户、状态、IP分页查询登录日志
      parameters:
      - description: 页码
        in: query
        name: pageNum
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 用户名
        in: query
        name: userName
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: IP地址
        in: query
        name: ipaddr
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取登录日志
      tags:
      - Monitor/LoginLog
  /v1/monitor/logs/login/{id}:
    delete:
      description: 根据ID删除登录日志
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 删除登录日志
      tags:
      - Monitor/LoginLog
    get:
      description: 根据ID查看登录日志
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 获取登录日志详情
      tags:
      - Monitor/LoginLog
  /v1/monitor/logs/login/{id}/unlock:
    post:
      description: 通过登录日志记录解除账号锁定
      parameters:
      - description: 日志ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resp.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resp.Response'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/resp.Response'
      security:
      - BearerAuth: []
      summary: 解锁账号
      tags:
      - Monitor/LoginLog
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"